Experimental udp_send_all functions.
Not entirely satisfied with the naming.
send_all in the standard library returns an error when all data cannot be sent.
The send_all methods here can block indefinitely when all data cannot be sent.

This could lead to some nasty bugs if the behaviour is assumed to align with
the read_all and write_all methods from std::io.